fix(digest): address PR #26 review feedback
- Break circular dependency: extract LLM_PROVIDER token to llm.tokens.ts
  and createLlmProvider factory to llm.factory.ts so llm.module.ts no
  longer imports from index.ts (which re-exports from llm.module.ts)
- Add 60s timeout to ChatOpenAI constructor
- Add test for JSON.stringify fallback when response content is not a string
- Mark OPENROUTER_API_KEY as required in .env.example
