+++ .env.example
@@ -19,7 +19,7 @@ X_SCOPES=tweet.read users.read like.read bookmark.read offline.access
 
 # ----- LLM (default: openrouter) -----
 LLM_PROVIDER=openrouter
-OPENROUTER_API_KEY=
+OPENROUTER_API_KEY=          # required
 OPENROUTER_MODEL=anthropic/claude-sonnet-4.5
 
 # ----- Article extractor (default: firecrawl) -----
