+++ src/digest/llm/llm.module.ts
@@ -1,16 +1,7 @@
 import { type DynamicModule, Module } from '@nestjs/common';
 import type { Env } from '../../config/env';
-import { createLlmProvider } from './index';
-
-/**
- * DI token for the `LlmProvider` port. Consumers inject the provider
- * via `@Inject(LLM_PROVIDER)` without referencing the concrete class.
- *
- * Follows the same naming convention as `X_SOURCE` in
- * `src/ingestion/ingestion.module.ts` and `X_OAUTH_CLIENT` in
- * `src/auth/auth.module.ts`.
- */
-export const LLM_PROVIDER = 'LlmProvider';
+import { createLlmProvider } from './llm.factory';
+import { LLM_PROVIDER } from './llm.tokens';
 
 /**
  * Wires the LLM provider into the NestJS DI container.
